requirements.txt
Django>=4.0,<5.0
psycopg2-binary>=2.9.6
elasticsearch-dsl>=7.4.0
line-bot-sdk